+++ CHANGELOG.rst
@@ -1,7 +1,8 @@
 0.1.20 (unreleased)
 ===================
 
-- Nothing changed yet.
+- Fields no longer have a default.
+- Adapter.extra_headers returns a dictionary instead of a list
 
 
 0.1.19 (2015-03-16)